\name{regression_analysis_plots}
\alias{regression_analysis_plots}
\title{Generate Regression Plots with Absolute Errors}
\description{
  This function generates scatter plots for the specified regression model
  (Linear Regression, SVM, KSVM, or Random Forest) along with the absolute error
  for each prediction.
}
\usage{
regression_analysis_plots(data, target_var, predictor_var, model_choice)
}
\arguments{
  \item{data}{A data frame containing the data for modeling.}
  \item{target_var}{The name of the target variable (dependent variable) in the
  data frame. It should be a column in the data.}
  \item{predictor_var}{The name of the predictor variable (independent variable)
  in the data frame. It should be a column in the data.}
  \item{model_choice}{A string specifying the model to plot. Must be one of "LM"
  (Linear Model), "SVM" (Support Vector Machine), "KSVM" (Kernel SVM), or "RF"
  (Random Forest).}
}
\value{
  A ggplot object representing the scatter plot with absolute errors.
}
\examples{
# Example: Regression analysis plots for 'Wind' vs 'Temp' using LM
regression_analysis_plots(airquality, target_var = "Wind", predictor_var = "Temp", model_choice = "LM")
}
